+++ aeon/testing/utils/data_gen/__init__.py
@@ -28,7 +28,6 @@ __all__ = [
     "_assert_correct_columns",
     "_assert_correct_pred_time_index",
     "make_annotation_problem",
-    "_convert_tsf_to_hierarchical",
     "_get_n_columns",
     "get_examples",
     "make_example_2d_unequal_length",
@@ -49,7 +48,6 @@ from aeon.testing.utils.data_gen._collection import (
     make_example_nested_dataframe,
     make_example_unequal_length,
 )
-from aeon.testing.utils.data_gen._data_generators import _convert_tsf_to_hierarchical
 from aeon.testing.utils.data_gen._series import (
     _make_index,
     make_forecasting_problem,
